refactor: CarName의 원시값 추출에 따른 로직 및 테스트코드 수정
