+++ src/main/java/racingcar/domain/Car.java
@@ -4,22 +4,13 @@ public class Car {
 
     public static final int START_POSITION = 0;
     public static final int MOVABLE_CONDITION = 4;
-    public static final int NAME_MAX_LENGTH = 5;
 
     private Position position;
-    private String name;
+    private CarName name;
 
     public Car(String name) {
-        validateNameLength(name);
-
-        this.name = name;
+        this.name = new CarName(name);
         this.position = new Position(START_POSITION);
-    }
-
-    private void validateNameLength(String name) {
-        if (name.length() >= NAME_MAX_LENGTH) {
-            throw new IllegalArgumentException("이름의 길이가 너무 깁니다. : " + name.length());
-        }
     }
 
     public void forwardByCondition(int condition) {
@@ -32,7 +23,7 @@ public class Car {
         return position;
     }
 
-    public String getName() {
+    public CarName getName() {
         return name;
     }
 }
